Module runa::math

The Math crate of the project

Structs

Align

Traits

MetricSpace

A structure meant to be passed as a generic parameter to other functions. it is meant to ensapsulate all the information related to

Functions

align_lettre_mot

Align a letter with a word in the optimal way. O(n) time

coupure

Calculate the optimal cutting point in sequence y for the corresponding cutting point |x|/2 in sequence x

cout_align

Calculate the cost of the alignment (x, y) passed as parameter

dist_1

Compute distance using a 2D table O(n^2) time O(n^2) memory

dist_2

Calculate the distance between two sequences O(n^2) time O(n) memory

dist_dp_full

Compute the 2D dynamic-programming table for the sequences x and y

dist_naif

Calculate distance between sequences x and y in the MetricSpace M O(exp(n)) time and memory

dist_naif_rec

auxiliary function for dist_naif

mot_gaps

Generate a word composed of gaps of length n

prog_dyn

Calculate the optimal alignment and the distance between the sequences at once

rm_gaps

Remove all gaps from the sequence

sol_1

Compute the optimal alignment using a 2D table O(n^2) time and memory

sol_1_tab

Same as sol_1 but you pass in the table manually

sol_2

Compute the alignement of two sequences in O(n^2) time and O((n+m)log n) memory

sol_2_ll

Auxiliary function for sol_2 O(n^2) time and O((n+m)log n) memory